refactor kv pairs
